Create About.py
Heart disease is a significant health concern worldwide, and the ability to predict the likelihood of its occurrence can be of great value for early intervention and prevention. In this project, we will develop a heart disease prediction model using Python. The model will utilize a dataset containing various attributes and medical measurements of patients to predict the presence or absence of heart disease.